+++ apx_sdk/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Added documentation to `apx_sdk::addresses` module.
+
 ## [0.22.0] - 2026-03-27
 
 ### Added